+++ facebook_auth_desktop/example/lib/main.dart
@@ -1,42 +1,65 @@
+import 'dart:convert';
+
 import 'package:flutter/material.dart';
 import 'package:facebook_auth_desktop/facebook_auth_desktop.dart';
 
 final plugin = FacebookAuthDesktopPlugin();
 
+String prettyPrint(Map json) {
+  JsonEncoder encoder = const JsonEncoder.withIndent('  ');
+  String pretty = encoder.convert(json);
+  return pretty;
+}
+
 void main() {
   plugin.webAndDesktopInitialize(
     appId: '1329834907365798',
     cookie: true,
-    version: 'v13.0',
+    version: 'v15.0',
     xfbml: true,
   );
   runApp(const MyApp());
 }
 
-class MyApp extends StatefulWidget {
+class MyApp extends StatelessWidget {
   const MyApp({Key? key}) : super(key: key);
 
   @override
-  State<MyApp> createState() => _MyAppState();
+  Widget build(BuildContext context) {
+    return const MaterialApp(
+      home: HomePage(),
+    );
+  }
 }
 
-class _MyAppState extends State<MyApp> {
-  void _signIn() {
-    plugin.login();
+class HomePage extends StatelessWidget {
+  const HomePage({Key? key}) : super(key: key);
+
+  void _signIn(BuildContext context) async {
+    final result = await plugin.login();
+    if (result.status.name == 'success') {
+      final userData = await plugin.getUserData();
+      showDialog(
+        context: context,
+        builder: (_) => AlertDialog(
+          content: Text(
+            prettyPrint(userData),
+          ),
+        ),
+      );
+    }
   }
 
   @override
   Widget build(BuildContext context) {
-    return MaterialApp(
-      home: Scaffold(
-        appBar: AppBar(
-          title: const Text('Plugin example app'),
-        ),
-        body: Center(
-          child: MaterialButton(
-            onPressed: _signIn,
-            child: const Text('sign in'),
-          ),
+    return Scaffold(
+      appBar: AppBar(
+        title: const Text('Plugin example app'),
+      ),
+      body: Center(
+        child: MaterialButton(
+          onPressed: () => _signIn(context),
+          child: const Text('sign in'),
         ),
       ),
     );