+++ facebook_auth_desktop/CHANGELOG.md
@@ -1,3 +1,5 @@
+## 0.0.6
+- fixed `long_lived_token` is null on macOs.
 ## 0.0.5
 - flutter_facebook_auth_platform_interface: ^4.1.1
 ## 0.0.4